Changing byte to int be able to spawn more than 255 bees

+++ Ported/CombatBees/Assets/Scripts/Components/Spawners/BeeSpawnerAuthoring.cs
@@ -6,5 +6,5 @@ public struct BeeSpawner : IComponentData
 {
 	public Entity BeePrefab_TeamA;
 	public Entity BeePrefab_TeamB;
-	public byte Count;
+	public int Count;
 }
